rename: open_webui.pipelines.schemas
