+++ open_webui/pipelines/main.py
@@ -17,7 +17,7 @@ from pydantic import BaseModel, ConfigDict
 from starlette.responses import Response, StreamingResponse
 
 from open_webui.pipelines.utils import get_last_user_message, stream_message_template
-from schemas import OpenAIChatCompletionForm
+from open_webui.pipelines.schemas import OpenAIChatCompletionForm
 
 PIPELINES = {}
