+++ open_webui/pipelines/schemas.py
